Add initial NoteDirectoryWatcher add-in for watching external changes to note files.
If you use an external process to modify you ".note" files (ie
DropBox) there is a significant risk that you will lose you note data.
Tomboy will overwrite the note on disk with its representation in
memory.

This patch will watch your ~/.tomboy directory for changes
and update Tomboy's in-memory copy of the note if the note changes.
This reduces the change of losing notes.