fixed GMime usage to support 2.2 and 2.4
Fixes for bug #565336

Removed unnecessary message.Subject header processing from
EvolutionNoteAddin.cs which was probably used to make sure
that the subject was in UTF-8 when using GMime versions older
than 2.2.7 (altho it'd also open up a new can of worms...).

Changed configure.in to prefer GMime# 2.4.x but to gracefully
fall back to 2.2.x >= 2.2.7.