Delay most work until after tray icon appears for 0.65 second startup
Refactor NoteManager to push its work to an Initialize method, and
provide Invoke methods that allow code to be executed only when the
NoteManager is fully initialized.

Use these Invoke methods to make clicking the tray icon and using global
keybindings safe.  RemoteControl work is on the way.

Also delay SyncManager and ApplicationAddin initialization.