Restructure RemoteControl and command line execution
One important thing to note is that we are no longer setting
MONO_DISABLE_SHM=1 (see bug #514434), which means that Tomboy
may write to disk every 40 seconds unless you upgrade to Mono 2.4.3
or later.

This change was made so that we can depend on a named mutex for
single-instance detection.