Do not perform a sync less than a minute after a note save occurs
This is basically a poor man's idle detection. Previously, even if you were
in the middle of typing in a note window, a sync might occur and interrupt
your work.  Attempting to catch all user events seemed like it would be a
messy way to implement idle detection, so instead we catch note save events
and if save occurs within the last minute of the autosync timer, we push
the timer back to a whole minute.

This is an incomplete solution in many ways, but is Good Enough for now.