[Setup] Improve Tomboy's presentation for Add/Remove programs
Add the Tomboy icon as program icon to the installer database.
Add the link to the Tomboy end-user site as help link.

Define the FileVersion for the resulting Tomboy.exe assembly and
automatically use that version as ProductVersion in the MSI table.
Drop the AssemblyInformationalVersion and base Defines.VERSION on
the assembly version again, reducing the number of lines to change
for a new windows version to exactly one.