[Windows] Add the missing resource "tomboy-note-clipboard-html.xsl" and clean up the msbuild file.
Add the resource file to the windows build (it was present in the makefiles already)
to fix bug #625510.

Clean up the references to resources and point them to the real location below data/
instead of the project root. Before this change a clean clone of Tomboy shows missing
resources in Visual Studio, the files are (unintuitively) copied to the right spot in
a pre-build command using a bunch of "copy" invocations.

Remove the post-build event and migrate it to a standard "AfterBuild" target, replacing
invocations of mkdir/copy with <MakeDir>/<Copy>. That way it works on every OS that
supports msbuild (or xbuild on mono).