[mac] Use hand cursor when hovering over links
MouseHandWatcher only changes the cursor to the hand cursor if Ctrl or
Shift is not pressed when hovering a link. That did not work on mac
because the modifiers obtained from GdkWindow had some modifiers set
although none of them were actually held down. We now use the
coordinates and modifiers from the event args.

https://bugzilla.gnome.org/show_bug.cgi?id=616317