Rewrite Trie.cs to fix bug 584362 and move to a cleaner code base
The previous implementation failed to track the start of a match correctly if
fail states lead to a match. This resulted in a wrong substring being returned
as "Key" and wrong indices to select the relevant substring in the haystack.

In addition the previous implementation created the failure graph on every
keyword insert. This patch does this once, after all note titles are inserted.
On my old, dusty laptop this improves startup performance by ~2.5 seconds..

Signed-off-by: Benjamin Podszun <benjamin.podszun@gmail.com>