Aimed at fixing the off-screen Search window on some multi-monitor configurations
We determine the number of the monitor upon saving the window position
and during the restoration we check if that's the same one or not. If
yes - we restore as-is (assuming that the setup didn't change and the
coordinates should be valid), if no - we restore to the center of the
screen.

Signed-off-by: Jared Jennings <jjennings@src.gnome.org>