Use dmcs if present, falling back on gmcs
This causes Tomboy to default to being compiled against the 4.0 profile if dmcs
is present in the system. This will fix the issue with Tomboy needing
GMCS=mono-csc or GMCS=dmcs in ./configure on Debian, Ubuntu, and other systems
that default to the 4.0 profile.

This should be harmless on systems that default to the 2.0 profile as 4.0
applications can use libraries compiled against the 2.0 profile without issue,
but not vice versa.

Signed-off-by: Chow Loong Jin <hyperair@debian.org>
Signed-off-by: Jared Jennings <jjennings@src.gnome.org>