Fixed installer asking to install gtk# when it's already installed
Some systems may have AllwaysInstallElevated property set to "1" and as a
result some of the crucial properties our installer uses are not passed
between the installation phases, because they're not listed as secure.

That causes the installer to think user doesn't have gtk# installed.

This patch marks all properties as secure explicitly and fixed that
behavior.

Signed-off-by: Jared Jennings <jared@jaredjennings.org>