Remove the Evolution add-in
This is because it relies on knowledge of Evolution internals and gConf settings,
anno 2004. Evolution has migrated to gSettings and has probably changed a lot internally,
and the way the add-in reproduced internals is enough to give me the screaming heebie
jeebies anyway. Did I mention that it also P\Invoked stuff?

...

NUKED FROM ORBIT