build: add ax_require_config macro copy to our code
This is to alleviate the problem with building the package on distros
with older autoconf-archive version, which does not include this macro.
Ubuntu 14 is one such distro.

aclocal will overwrite our version if the OS has a newer version,
as determined by the "serial" line in the macro.

We will not track other files in m4 directory though,
as the assumption is that such cases should be rare.

Closes #26.

Signed-off-by: Alex Tereschenko <frozen.and.blue@gmail.com>